+++ common/v2/database/index.ts
@@ -1,1 +1,2 @@
 export * from './v1.0.0';
+export { getCurrentDBConfig, getData } from './versions';
