+++ common/v2/database/v1.0.0/index.ts
@@ -1,6 +1,7 @@
 import { SCHEMA_BASE } from './schema';
-import { createSchema } from './migration';
+import { createDefaultValues } from './generateDefaultValues';
 
 export { addDevSeedToSchema } from './devSeed';
 export { removeSeedDataFromSchema } from './removeSeed';
-export const SCHEMA_DEFAULT = createSchema(SCHEMA_BASE);
+export { migrate } from './migration';
+export const SCHEMA_DEFAULT = createDefaultValues(SCHEMA_BASE);
