--- common/v2/services/Store/DataManager/constants.ts
@@ -1,4 +1,0 @@
-// The name of the key in LocalStorage used for persistence.
-export const LOCALSTORAGE_KEY = 'MyCryptoStorage';
-
-export const ENCRYPTED_STORAGE_KEY = 'MyCryptoEncrypted';
